requirements.txt
defusedxml==0.4.1
Django==1.9.6
django-rest-swagger==0.3.7
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
PyYAML==3.11
wheel==0.24.0
